Adding NIH CSL for mendeley
Adds a DOI and PMCID for the document. PMIDs will be used if PMCID is not provided.

To add PMCID in mendeley, go to "Notes" tab and add the id in the first two sentences of the notes section using the format:

{:PMCID: 0000000} - replace the 0000000 with actual PMCID.

Any text entered in a new line / outside of the brackets will be ignored in the bibliography.